pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ransomemu"
version = "0.1.0"
description = "Ransomware emulation framework for Exegol — Purple/Red Team network resilience testing"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "RansomEmu Team"},
]

dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "langchain>=0.3",
    "langchain-ollama>=0.3",
    "langchain-core>=0.3",
    "impacket>=0.12",
    "pywinrm>=0.4",
    "httpx>=0.27",
    "jinja2>=3.1",
    "networkx>=3.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "mypy>=1.11",
    "respx>=0.21",
]

[project.scripts]
ransomemu = "ransomemu.cli:cli"

[tool.setuptools.packages.find]
include = ["ransomemu*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B"]
ignore = ["S603", "S607"]  # subprocess security — intentional in this tool

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
